Remove show_undo_requested prop and replace with a hook.
