+++ src/views/Game/GameHooks.ts
@@ -1,0 +1,46 @@
+/*
+ * Copyright (C) 2012-2022  Online-Go.com
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import * as React from "react";
+import { GobanCore } from "goban";
+import { game_control } from "./game_control";
+
+/** React hook that returns true if an undo was requested on the current move */
+export function useUndoRequested(goban: GobanCore): boolean {
+    const [show_undo_requested, setShowUndoRequested] = React.useState(
+        goban.engine.undo_requested === goban.engine.last_official_move.move_number,
+    );
+    React.useEffect(() => {
+        const syncShowUndoRequested = () => {
+            if (game_control.in_pushed_analysis) {
+                return;
+            }
+
+            setShowUndoRequested(
+                goban.engine.undo_requested === goban.engine.last_official_move.move_number &&
+                    goban.engine.undo_requested === goban.engine.cur_move.move_number,
+            );
+        };
+        syncShowUndoRequested();
+
+        goban.on("load", syncShowUndoRequested);
+        goban.on("undo_requested", syncShowUndoRequested);
+        goban.on("last_official_move", syncShowUndoRequested);
+    });
+
+    return show_undo_requested;
+}
